javascript checks up on start, shows caution
popup now appears while server is starting, or if there are errors, and comes back if server changes to standby.
changed to plug in *before* the routes are evaluated, so we don't load the wrong data.

expands rest api to return several pieces of information to simplify gui checks.

simplifies code in rest filters to block requests not compatible with server state, including checking for upness there.